// vibeui/src/app/api/model/itemAliases.ts
/**
 * CRISTAL-iSE Item Rest API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ItemAliases { 
    uuid: string;
    name?: string;
    domainPaths?: Array<string>;
    error?: string;
}