// vibeui/src/app/api/model/pathData.ts
/**
 * CRISTAL-iSE Item Rest API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PathData { 
    name: string;
    type: PathData.TypeEnum;
    url: string;
    path: string;
    uuid?: string;
    hasJoblist?: boolean;
}
export namespace PathData {
    export const TypeEnum = {
        Domain: 'domain',
        Item: 'item',
        Agent: 'agent',
        Role: 'role',
        NA: 'n/a'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}